+++ src/pages/annotate/RelationEditor/RelationshipBrowser/RelationshipBrowserSuggestion.tsx
@@ -1,6 +1,5 @@
-import { useDataModel } from '@/store';
 import { useMemo } from 'react';
-import { RelationshipSearchResult } from './useRelationshipSearch';
+import { RelationshipSearchResult, useDataModel } from '@/store';
 
 interface RelationshipBrowserSuggestionProps {
