+++ src/store/hooks/index.ts
@@ -6,4 +6,5 @@ export * from './useFolderMetadata';
 export * from './useImageMetadata';
 export * from './useImages';
 export * from './useImageSnippet';
+export * from './useRelationshipSearch';
 export * from './useStore';
